fix: set minimum version for graphviz
The Alpine repo provides only the latest version of packages. The
previous release is purged when a new one is uploaded, causing docker
builds to fail. To work around this, I'm setting a minimum version to
the minor version. It will still fail if graphviz decides to release
v9.1, but that's possibly a breaking change that we should re-test
anyway. See the official documentation for syntax.
https://wiki.alpinelinux.org/wiki/Alpine_Package_Keeper#Advanced_APK_Usage